requirements.txt
fastapi
Jinja2
uvicorn
gunicorn
jinja2
gspread
psycopg2
SQLAlchemy
bcrypt
qrcode
pillow